requirements.txt
asgiref
certifi
chardet
charset-normalizer
dj-database-url
Django
django-cors-headers
django-heroku
djangorestframework
gunicorn
idna
npm
optional-django
Pillow
psycopg2
psycopg2-binary
python-dotenv
pytz
requests
sqlparse
urllib3
whitenoise
